pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "benthoscan"
version = "0.0.1"
description = "A small API for 3D reconstruction and registration based on images"
authors = ["Martin Kvisvik Larsen <martin.kvisvik.larsen@hotmail.com>"]
readme = "README.md"
packages = [{include = "benthoscan", from="src"}]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.11"
arrow = "^1.3.0"
h5py = "^3.11.0"
loguru = "^0.7.2"
numpy = "^1.26.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
result = "^0.16.1"
toml = "^0.10.2"
tqdm = "^4.66.2"
open3d = "^0.18.0"
metashape = {path = "vendor/Metashape-2.1.0-cp37.cp38.cp39.cp310.cp311-abi3-linux_x86_64.whl"}
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"

[tool.poetry.scripts]
create = "benthoscan.runtime.cli:main"
summarize = "benthoscan.runtime.cli:main"
